Initial conversion of the definition to a LWRP
Uses a more expected format for the actions (enable/disable)
Adds why-run support

This IS a breaking change.  The current functionality requires passing an enabled param to enabled and if that's not present it removes the rule.  With the LWRP we'll default to enable and have an action for disable.  This breaks existing implementations but better aligns with how LWRPs work in chef.