Fix rebuild-iptables for IPv6
rebuild-iptables fails on restoring the generated /etc/sysconfig/ip6tables rule
file as it tried to use iptables-restore to load the rules.

Changed to use the ipt variable to select the relevant command to use and added
a better test case of a rule to allow DHCPv6 as it uses IPv6 addresses
(and more importantly prefix length) which will fail if the wrong command is
used.

Signed-off-by: Ben Hughes <bmhughes@bmhughes.co.uk>