Refactor rule resource to use accumulator
The rule (and by virtue rule6) resource has been majorly refactored to
use an accumulated template resource to generate the persistent rules
file for both iptables and ip6tables in a more 'chefy' way.

The generator script and ip(6)tables.d directories are no longer needed
and have been removed from the cookbook.

The new resources can be used in a similar way to the previous ones by
feeding it a full iptables rule statement, or by settings the resource
properties and allowing it to generate the final statement.

In addition, a chain(6) resource has been added to allow the 'chefy' generation of
iptables chains for rules to be assigned.

Signed-off-by: Ben Hughes <bmhughes@bmhughes.co.uk>